Add files via upload
This code is basically coded in python with matplotlib and numpy module . This is a research based code for my project , In which we can predict the GRAIN GROWTH .
Grain growth refers to an increase in the size of crystallites (grains) in a material at high temperatures.

Grain growth is due to an alteration of the internal energy of the molecular structure of a given metal, ceramic or mineral. Grain growth is typical in metals that have undergone annealing and other processes of surface modification. Due to the destabilization of the molecular structures of metals due to grain growth, the probability of crevice, intergranular and grain boundary corrosion are increased.